Changed variables in struct, finishing FCFS, and more.
- CRITICAL CHANGE in struct: they MUST ALL be pointers, except for
length.
-> Done changes in MyScheduler.c, Utils_main.c, and in Utils.c as well.
- debugMode is unsigned SHORT int.
KNOWN ISSUES:

- missing test with Duration time = 0
- missing test with mixed arrivals time.

+++ SchedSIM_progetto_2020/Utils_main.h
@@ -48,7 +48,7 @@ short int selectMode(unsigned short int debugMode);
 // Funzione che scansiona rapidamente il file, contando quanti processi ci sono.
 int countProcesses(char* buffer, unsigned short int debugMode);
 // Funzione che preprocessa il buffer, sostituendo i token separatori in spazi.
-int preProcess(char* buffer, unsigned int debugMode);
+int preProcess(char* buffer, unsigned short int debugMode);
 // Funzione che memorizza il contenuto del buffer nei vari array.
 int importProcesses(char* buffer, struct processesData processes, unsigned short int debugMode);
 // Funzione che libera lo spazio allocato dinamicamente per l'array dei nomi di processo.
